+++ dali/operators/geometry/affine_transforms/transform_base_op.h
@@ -33,8 +33,25 @@ namespace dali {
 template <int... values>
 using dims = std::integer_sequence<int, values...>;
 
-template <typename T, int mat_dim>
-using affine_mat_t = mat<mat_dim, mat_dim, T>;
+template <typename T, int ndim>
+using affine_mat_t = mat<ndim+1, ndim+1, T>;
+
+template <typename Value, typename Callable, typename... Args>
+void value_switch(Value value, std::integer_sequence<Value>,
+                  Callable &&callable, Args&&... args) {
+  throw std::invalid_argument(make_string("Unsupported parameter value: ", value));
+}
+template <typename Value, typename Callable, typename... Args, Value first, Value... values>
+void value_switch(Value value, std::integer_sequence<Value, first, values...>,
+                  Callable &&callable, Args&&... args) {
+  if (value == first) {
+    callable(std::integral_constant<Value, first>(), std::forward<Args>(args)...);
+  } else {
+    value_switch(value, std::integer_sequence<Value, values...>(),
+                 std::forward<Callable>(callable), std::forward<Args>(args)...);
+  }
+}
+
 
 /**
  * @brief Base CRTP class for affine transform generators.
@@ -44,9 +61,10 @@ using affine_mat_t = mat<mat_dim, mat_dim, T>;
 template <typename Backend, typename TransformImpl>
 class TransformBaseOp : public Operator<Backend> {
  public:
-  explicit TransformBaseOp(const OpSpec &spec) :
-      Operator<Backend>(spec),
-      reverse_order_(spec.GetArgument<bool>("reverse_order")) {
+  explicit TransformBaseOp(const OpSpec &spec)
+      : Operator<Backend>(spec),
+        spec_(spec),
+        nsamples_(spec.GetArgument<int>("batch_size")) {
     matrix_data_.set_pinned(false);
     matrix_data_.set_type(TypeTable::GetTypeInfo(dtype_));
   }
@@ -67,7 +85,7 @@ class TransformBaseOp : public Operator<Backend> {
   }
 
   bool SetupImpl(std::vector<OutputDesc> &output_descs, const workspace_t<Backend> &ws) override {
-    has_input_ = ws.NumInput() > 0;
+    has_input_ = ws.template NumInput() > 0;
     if (has_input_) {
       auto &input = ws.template InputRef<Backend>(0);
       const auto &shape = input.shape();
@@ -82,8 +100,6 @@ class TransformBaseOp : public Operator<Backend> {
           "The input, if provided, is expected to be a 2D tensor with dimensions "
           "(ndim, ndim+1) representing an affine transform. Got: ", shape));
       nsamples_ = shape.num_samples();
-    } else {
-      nsamples_ = spec_.template GetArgument<int>("batch_size");
     }
 
     This().ProcessArgs(spec_, ws);
@@ -95,27 +111,16 @@ class TransformBaseOp : public Operator<Backend> {
     return true;
   }
 
-  template <typename T>
-  void RunImplTyped(workspace_t<Backend> &ws, dims<>) {
-    DALI_FAIL(make_string("Unsupported number of dimensions ", ndim_));
-  }
-
-  template <typename T, int ndim, int... ndims>
-  void RunImplTyped(workspace_t<Backend> &ws, dims<ndim, ndims...>) {
-    if (ndim_ != ndim) {
-      RunImplTyped<T>(ws, dims<ndims...>());
-      return;
-    }
-
-    constexpr int mat_dim = ndim + 1;
+  template <typename T, int ndim>
+  void RunImplTyped(std::integral_constant<int, ndim>, workspace_t<Backend> &ws) {
     auto &out = ws.template OutputRef<Backend>(0);
     out.SetLayout({});  // no layout
 
     int64_t num_mats = This().IsConstantTransform() ? 1 : nsamples_;
-    matrix_data_.Resize({num_mats * static_cast<int64_t>(sizeof(affine_mat_t<T, mat_dim>))});
-    span<affine_mat_t<T, mat_dim>> matrices{
-        reinterpret_cast<affine_mat_t<T, mat_dim> *>(matrix_data_.mutable_data<T>()), num_mats};
-    This().DefineTransforms(matrices);
+    matrix_data_.Resize({num_mats * static_cast<int64_t>(sizeof(affine_mat_t<T, ndim>))});
+    span<affine_mat_t<T, ndim>> matrices{
+        reinterpret_cast<affine_mat_t<T, ndim> *>(matrix_data_.mutable_data<T>()), num_mats};
+    This().template DefineTransforms<T, ndim>(matrices);
 
     auto out_view = view<T>(out);
     if (has_input_) {
@@ -123,12 +128,12 @@ class TransformBaseOp : public Operator<Backend> {
       auto in_view = view<T>(in);
       for (int i = 0; i < nsamples_; i++) {
         int mat_idx = num_mats == 1 ? 0 : i;
-        ApplyTransform(out_view[i].data, in_view[i].data, matrices[mat_idx]);
+        ApplyTransform<T, ndim>(out_view[i].data, in_view[i].data, matrices[mat_idx]);
       }
     } else {
       for (int i = 0; i < nsamples_; i++) {
         int mat_idx = num_mats == 1 ? 0 : i;
-        ApplyTransform(out_view[i].data, matrices[mat_idx]);
+        ApplyTransform<T, ndim>(out_view[i].data, matrices[mat_idx]);
       }
     }
   }
@@ -136,7 +141,9 @@ class TransformBaseOp : public Operator<Backend> {
   void RunImpl(workspace_t<Backend> &ws) override {
     TYPE_SWITCH(dtype_, type2id, T, TRANSFORM_INPUT_TYPES, (
       using SupportedDims = typename TransformImpl::SupportedDims;
-      RunImplTyped<T>(ws, SupportedDims());
+      value_switch(ndim_, SupportedDims(),
+        [this](auto &&... args) { RunImplTyped<T>(args...); },
+        ws);
     ), DALI_FAIL(make_string("Unsupported data type: ", dtype_)));  // NOLINT
   }
 
@@ -150,9 +157,8 @@ class TransformBaseOp : public Operator<Backend> {
   }
 
  private:
-  template <typename T, int mat_dim>
-  void ApplyTransform(T *transform_out, const affine_mat_t<T, mat_dim> &M) {
-    constexpr int ndim = mat_dim - 1;
+  template <typename T, int ndim>
+  void ApplyTransform(T *transform_out, const affine_mat_t<T, ndim> &M) {
     for (int i = 0, k = 0; i < ndim; i++) {
       for (int j = 0; j < ndim + 1; j++, k++) {
         transform_out[k] = M(i, j);
@@ -160,10 +166,9 @@ class TransformBaseOp : public Operator<Backend> {
     }
   }
 
-  template <typename T, int mat_dim>
-  void ApplyTransform(T *transform_out, const T* transform_in, const affine_mat_t<T, mat_dim> &M) {
-    constexpr int ndim = mat_dim - 1;
-    auto mat_in = affine_mat_t<T, mat_dim>::identity();
+  template <typename T, int ndim>
+  void ApplyTransform(T *transform_out, const T* transform_in, const affine_mat_t<T, ndim> &M) {
+    auto mat_in = affine_mat_t<T, ndim>::identity();
     for (int i = 0, k = 0; i < ndim; i++) {
       for (int j = 0; j < ndim + 1; j++, k++) {
         mat_in(i, j) = transform_in[k];
@@ -171,7 +176,7 @@ class TransformBaseOp : public Operator<Backend> {
     }
 
     // matrix multiplication
-    auto mat_out = reverse_order_ ? mat_in * M : M * mat_in;
+    auto mat_out = M * mat_in;
 
     for (int i = 0, k = 0; i < ndim; i++) {
       for (int j = 0; j < ndim + 1; j++, k++) {
@@ -181,15 +186,13 @@ class TransformBaseOp : public Operator<Backend> {
   }
 
  protected:
+  const OpSpec& spec_;
   DALIDataType dtype_ = DALI_FLOAT;
   int ndim_ = -1;  // will be inferred from the arguments or the input
   int nsamples_ = -1;
   bool has_input_ = false;
-  bool reverse_order_ = false;
 
   Tensor<CPUBackend> matrix_data_;
-
-  using Operator<Backend>::spec_;
 };
