--- dali/test/python/test_operator_transform.py
@@ -1,33 +1,0 @@
-# Copyright (c) 2020, NVIDIA CORPORATION. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from nvidia.dali.pipeline import Pipeline
-import nvidia.dali.ops as ops
-import nvidia.dali.types as types
-import nvidia.dali.fn as fn
-import numpy as np
-import os
-
-def test_translate_transform_op(batch_size=1, num_threads=4, device_id=0):
-    pipe = Pipeline(batch_size=batch_size, num_threads=num_threads, device_id=device_id)
-    with pipe:
-        T1 = fn.translate_transform(device='cpu', offset=(2, 3))
-        pipe.set_outputs(T1)
-    pipe.build()
-    outs = pipe.run()
-    for idx in range(batch_size):
-        T1 = outs[0].at(idx)
-        print(T1)
-
-test_translate_transform_op()
